Remove all warnings during cargo test.
Primarily replace uses of the deprecated ::std::thread::sleep_ms with the util
test function sleep_ms.

Also did some small cleanup of imports.

+++ test/test.rs
@@ -54,7 +54,8 @@ mod ports {
     }
 }
 
-pub fn sleep_ms(ms: usize) {
+#[allow(deprecated)]
+pub fn sleep_ms(ms: u64) {
     use std::thread;
     thread::sleep_ms(ms as u32);
 }
